+++ android/versioned-abis/expoview-abi44_0_0/src/main/java/abi44_0_0/expo/modules/notifications/service/NotificationsService.kt
@@ -7,10 +7,6 @@ import android.content.Context
 import android.content.Intent
 import android.content.pm.ActivityInfo
 import android.net.Uri
-import android.os.Bundle
-import android.os.Parcel
-import android.os.Parcelable
-import android.os.ResultReceiver
 import android.util.Log
 import androidx.core.app.RemoteInput
 import expo.modules.notifications.notifications.model.Notification
@@ -29,6 +25,7 @@ import abi44_0_0.expo.modules.notifications.service.interfaces.CategoriesDelegat
 import abi44_0_0.expo.modules.notifications.service.interfaces.HandlingDelegate
 import abi44_0_0.expo.modules.notifications.service.interfaces.PresentationDelegate
 import abi44_0_0.expo.modules.notifications.service.interfaces.SchedulingDelegate
+import android.os.*
 import kotlin.concurrent.thread
 
 /**
@@ -424,11 +421,13 @@ open class NotificationsService : BroadcastReceiver() {
         intent.putExtra(IDENTIFIER_KEY, identifier)
       }
 
+      // We're defaulting to the behaviour prior API 31 (mutable) even though Android recommends immutability
+      val mutableFlag = if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.S) PendingIntent.FLAG_MUTABLE else 0
       return PendingIntent.getBroadcast(
         context,
         intent.component?.className?.hashCode() ?: NotificationsService::class.java.hashCode(),
         intent,
-        PendingIntent.FLAG_UPDATE_CURRENT
+        PendingIntent.FLAG_UPDATE_CURRENT or mutableFlag
       )
     }
 
@@ -458,11 +457,13 @@ open class NotificationsService : BroadcastReceiver() {
         intent.putExtra(NOTIFICATION_ACTION_KEY, action as Parcelable)
       }
 
+      // We're defaulting to the behaviour prior API 31 (mutable) even though Android recommends immutability
+      val mutableFlag = if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.S) PendingIntent.FLAG_MUTABLE else 0
       return PendingIntent.getBroadcast(
         context,
         intent.component?.className?.hashCode() ?: NotificationsService::class.java.hashCode(),
         intent,
-        PendingIntent.FLAG_UPDATE_CURRENT
+        PendingIntent.FLAG_UPDATE_CURRENT or mutableFlag
       )
     }